[FIX] #98 fixed funding detailpage ui
